[vulkan] Light culler binds depth texture. Object naming and warning fixes

+++ Engine/Video/Vulkan/Texture2DVulkan.cpp
@@ -98,8 +98,8 @@ void ae3d::Texture2D::Load( const FileSystem::FileContentsData& fileContents, Te
 
     if (static_cast<int>(GfxDeviceGlobal::properties.limits.maxSamplerAnisotropy) < GetFloatAnisotropy( anisotropy ))
     {
-        System::Print( "%s is using too big anisotropy (%f), max supported is %fx.\n", fileContents.path.c_str(), GetFloatAnisotropy( anisotropy ),
-            GfxDeviceGlobal::properties.limits.maxSamplerAnisotropy );
+        System::Print( "%s is using too big anisotropy (%f), max supported is %fx.\n", fileContents.path.c_str(), (double)GetFloatAnisotropy( anisotropy ),
+                       (double)GfxDeviceGlobal::properties.limits.maxSamplerAnisotropy );
         anisotropy = Anisotropy::k4;
     }
